refactor : 프로젝트 전체 구조를 리팩토링한다
- 각 클래스의 책임을 명확하게 하기 위해 클래스의 패키지를 분리
