+++ src/main/java/lotto/domain/Purchase.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public class Purchase {
     private final int amount;
