+++ src/test/java/lotto/BonusNumberTest.java
@@ -2,6 +2,8 @@ package lotto;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
+import lotto.domain.BonusNumber;
+import lotto.domain.WinningNumbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
