+++ src/test/java/lotto/PurchaseTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.Purchase;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
