+++ src/test/java/lotto/WinningNumbersTest.java
@@ -2,6 +2,7 @@ package lotto;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
+import lotto.domain.WinningNumbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
